comps/dataprep/pgvector/langchain/requirements.txt
beautifulsoup4
cairosvg
docarray[full]
docx2txt
easyocr
fastapi
html2text
huggingface_hub
langchain
langchain-community
langchain-text-splitters
markdown
numpy
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
pandas
pgvector==0.2.5
Pillow
prometheus-fastapi-instrumentator
psycopg2-binary
pymupdf
pyspark
pytesseract
python-docx
python-multipart
python-pptx
sentence_transformers
shortuuid
tiktoken
unstructured[all-docs]==0.15.7
uvicorn
